requirements.txt
crewai 
python-dotenv 
langchain_google_genai
langchain_community 
streamlit